next fix for warnings

+++ capi/ecmdStructs.C
@@ -4760,7 +4760,7 @@ void  ecmdIndexVectorEntry::printStruct() {
         printf("\tIndex: %d\n", index);
         printf("\trc: 0x%08x\n", rc);
 
-        printf("\tNumber of elements in buffer vector: %ld\n", buffer.size());
+        printf("\tNumber of elements in buffer vector: %lu\n", (unsigned long)buffer.size());
 
 }
 #endif  // end of REMOVE_SIM
